Added code to update sensor values in the vehicle update function.
Fixed case where intensity would be over 100 for distances between 0 and 1.

+++ Braitenberg Vehicle/Braitenberg Vehicle/Point.cpp
@@ -7,8 +7,8 @@ Point::Point(int _x, int _y) {
 }
 
 double Point::distance(Point _b) {
-	double dx = x - _b.x;
-	double dy = y - _b.y;
+	double dx = double(x - _b.x);
+	double dy = double(y - _b.y);
 	double result = sqrt(pow(dx, 2) + pow(dy, 2));
 	return result;
 }
